% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaleResids.R
\name{backscaleResids}
\alias{backscaleResids}
\title{Backscale residuals}
\usage{
backscaleResids(scaledResids, means, model)
}
\arguments{
\item{scaledResids}{scaled residuals}

\item{means}{a vector of means}

\item{model}{coefficients of a fitted mean-variance trend}
}
\description{
Backscale residuals
}
