% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootConfInt.R
\name{bootConfInt}
\alias{bootConfInt}
\title{Obtain confidence intervals for the raw effect sizes}
\usage{
bootConfInt(
  Total,
  idUnique,
  bootStraps,
  transforms,
  respS,
  B.B,
  method,
  CP,
  reps,
  n1,
  cutoff,
  R,
  fitResult,
  ...
)
}
\description{
Obtain confidence intervals for the raw effect sizes
}
