% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformations.R
\name{boxcox.transformation}
\alias{boxcox.transformation}
\title{Apply two-parameter Box-Cox transformation}
\usage{
boxcox.transformation(y, lambda, alpha = 0)
}
\arguments{
\item{y}{Numeric vector}

\item{lambda}{Power parameter in power transform}

\item{alpha}{Shift paramater in 2-parameter power transform. Defaults to
\code{0} which implies a 1-parameter Box-Cox transform.}
}
\value{
Power-transformed data
}
\description{
Apply two-parameter Box-Cox transformation
}
