% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MarginalFit-methods.R
\name{coef.MarginalFit}
\alias{coef.MarginalFit}
\title{Coefficients from marginal model estimation}
\usage{
\method{coef}{MarginalFit}(object, ...)
}
\arguments{
\item{object}{Output of \code{\link{fitMarginals}} function}

\item{...}{Further arguments}
}
\description{
Coefficients from marginal model estimation
}
