% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{col2hex}
\alias{col2hex}
\title{R color to RGB (red/green/blue) conversion.}
\usage{
col2hex(cname, alpha = FALSE)
}
\arguments{
\item{cname}{vector of any of the three kinds of R color specifications, i.e., either a color name (as listed by \code{\link{colors}}()), a hexadecimal string of the form "#rrggbb" or "#rrggbbaa" (see \code{\link{rgb}}), or a positive integer i meaning \code{\link{palette}}()[i].}

\item{alpha}{logical value indicating whether the alpha channel (opacity) values should be returned.}
}
\description{
R color to RGB (red/green/blue) conversion.
}
