% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ResponseSurface-methods.R
\name{contour.ResponseSurface}
\alias{contour.ResponseSurface}
\title{Method for plotting of contours based on maxR statistics}
\usage{
\method{contour}{ResponseSurface}(x, colorBy = "maxR", ...)
}
\arguments{
\item{x}{Output of \code{\link{fitSurface}}}

\item{colorBy}{String indicating the characteristic to use for coloring ("maxR" or "effect-size"). By default, "maxR".}

\item{...}{Further parameters passed to \code{\link{plot.maxR}}}
}
\description{
Method for plotting of contours based on maxR statistics
}
