% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MarginalFit-methods.R
\name{df.residual.MarginalFit}
\alias{df.residual.MarginalFit}
\title{Residual degrees of freedom in marginal model estimation}
\usage{
\method{df.residual}{MarginalFit}(object, ...)
}
\arguments{
\item{object}{Output of \code{\link{fitMarginals}} function}

\item{...}{Further arguments}
}
\description{
Residual degrees of freedom in marginal model estimation
}
