% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{directAntivirals}
\alias{directAntivirals}
\title{Partial data with combination experiments of direct-acting antivirals}
\format{A data frame with 3520 rows and 6 variables:
\itemize{
  \item experiment: ID of experiment (1-11)
  \item cpd1: name of the first compound (4 different compounds)
  \item cpd2: name of the second compound (11 different compounds)
  \item effect: observed effect (cell count)
  \item d1: dose of the first compound
  \item d2: dose of the second compound
}}
\description{
A dataset containing 11 combination experiments of direct-acting antivirals.
}
