% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MarginalFit-methods.R
\name{fitted.MarginalFit}
\alias{fitted.MarginalFit}
\title{Compute fitted values from monotherapy estimation}
\usage{
\method{fitted}{MarginalFit}(object, ...)
}
\arguments{
\item{object}{Output of \code{\link{fitMarginals}} function}

\item{...}{Further arguments}
}
\description{
Compute fitted values from monotherapy estimation
}
