% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ResponseSurface-methods.R
\name{fitted.ResponseSurface}
\alias{fitted.ResponseSurface}
\title{Predicted values of the response surface according to the given null model}
\usage{
\method{fitted}{ResponseSurface}(object, ...)
}
\arguments{
\item{object}{Output of \code{\link{fitSurface}}}

\item{...}{Further parameters}
}
\description{
Predicted values of the response surface according to the given null model
}
