% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformations.R
\name{get.abs_tval}
\alias{get.abs_tval}
\title{Return absolute t-value, used in optimization call in
\code{\link{optim.boxcox}}}
\usage{
get.abs_tval(value, fac, lambda, zero.add2 = 0)
}
\arguments{
\item{value}{data}

\item{fac}{factor}

\item{lambda}{box-cox parameter}

\item{zero.add2}{2nd box-cox parameter}
}
\description{
Return absolute t-value, used in optimization call in
\code{\link{optim.boxcox}}
}
