% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformations.R
\name{get.summ.data}
\alias{get.summ.data}
\title{Summarize data by factor}
\usage{
get.summ.data(value, fac)
}
\arguments{
\item{value}{data to sumamrize}

\item{fac}{factor to summarize by}
}
\description{
Summarize data by factor
}
