% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{getCP}
\alias{getCP}
\title{Estimate CP matrix from bootstraps}
\usage{
getCP(bootStraps, null_model, transforms, sigma0, doseGrid)
}
\arguments{
\item{bootStraps}{the bootstraps carried out already}

\item{null_model}{Specified null model for the expected response surface.
Currently, allowed options are \code{"loewe"} for generalized Loewe model,
\code{"hsa"} for Highest Single Agent model, \code{"bliss"} for Bliss additivity,
and \code{"loewe2"} for the alternative Loewe generalization.}

\item{transforms}{Transformation functions. If non-null, \code{transforms} is
a list containing 5 elements, namely biological and power transformations
along with their inverse functions and \code{compositeArgs} which is a list
with argument values shared across the 4 functions. See vignette for more
information.}

\item{sigma0}{standard deviation of the null model on the real data}

\item{doseGrid}{a grid of dose combinations}
}
\value{
Estimated CP matrix
}
\description{
This function is generally called from within \code{\link{fitSurface}}.
}
