% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getd1d2.R
\name{getd1d2}
\alias{getd1d2}
\title{A function to get the d1d2 identifier}
\usage{
getd1d2(dat)
}
\arguments{
\item{dat}{the data frame containing d1 and d2 entries}
}
\value{
a vector of d1d2 identifiers
}
\description{
A function to get the d1d2 identifier
}
