% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialMarginal.R
\name{GetStartGuess}
\alias{GetStartGuess}
\title{Estimate initial values for dose-response curve fit}
\usage{
GetStartGuess(df, transforms = NULL)
}
\arguments{
\item{df}{Dose-response dataframe containing \code{"dose"} and
\code{"effect"} columns}

\item{transforms}{Transformation functions. If non-null, \code{transforms} is
a list containing 5 elements, namely biological and power transformations
along with their inverse functions and \code{compositeArgs} which is a list
with argument values shared across the 4 functions. See vignette for more
information.}
}
\description{
Estimate initial values for dose-response curve fit
}
