% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsa.R
\name{hsa}
\alias{hsa}
\title{Highest Single Agent model}
\usage{
hsa(doseInput, parmInput, ...)
}
\arguments{
\item{doseInput}{Dose-response dataframe containing \code{"d1"} and
\code{"d2"} columns}

\item{parmInput}{Numeric vector or list with appropriately named
parameter inputs. Typically, it will be coefficients from a
\code{MarginalFit} object.}

\item{...}{Further arguments that are currently unused}
}
\description{
This function returns response levels for when these are based on
Highest Single Agent (HSA) model.
}
