% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isobologram.R
\name{isobologram}
\alias{isobologram}
\title{Isobologram of the response surface predicted by the null model}
\usage{
isobologram(x, grid.len = 100, logScale = TRUE, ...)
}
\arguments{
\item{x}{Output of \code{\link{fitSurface}}}

\item{grid.len}{Number of concentrations to plot for each compound in the
contour plot. An evenly spaced grid of doses will be generated for each
compound given its respective observed minimum and maximum doses. Note that
\code{grid.len^2} computations will be needed later so this number should
stay reasonably low.}

\item{logScale}{If \code{logScale = TRUE}, then grid of doses is evenly
spaced in the logarithmic scale.}

\item{...}{Further parameters that are not used at this moment.}
}
\description{
If transformation functions are used, then the isobologram response levels
will be plotted on the transformed scale.
}
