% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelVar.R
\name{modelVar}
\alias{modelVar}
\title{Calculate model variance, assuming variance increases linearly with mean}
\usage{
modelVar(dat_off)
}
\arguments{
\item{dat_off}{off-axis points data}
}
\value{
the predicted model variance
}
\description{
Calculate model variance, assuming variance increases linearly with mean
}
