% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confInt-functions.R
\name{plot.BIGLconfInt}
\alias{plot.BIGLconfInt}
\title{Plot confidence intervals in a contour plot}
\usage{
\method{plot}{BIGLconfInt}(x, ...)
}
\arguments{
\item{x}{Output from \code{\link{BIGLconfInt}}}

\item{...}{Further arguments}
}
\description{
Plot confidence intervals in a contour plot
}
