% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effectSize-functions.R
\name{plot.effect-size}
\alias{plot.effect-size}
\title{Plot of effect-size object}
\usage{
\method{plot}{`effect-size`}(
  x,
  main = "Contour plot for effect size",
  xlab = "Dose (Compound 1)",
  ylab = "Dose (Compound 2)",
  colorPalette,
  logScale = TRUE,
  zTransform = function(z) {
     z
 },
  digits,
  digitsFunc,
  ...
)
}
\arguments{
\item{x}{Object of class \code{effect-size}.}

\item{main}{The main title (on top) using font, size (character
    expansion) and color \code{par(c("font.main", "cex.main", "col.main"))}.}

\item{xlab}{X axis label using font, size and color 
    \code{par(c("font.lab", "cex.lab", "col.lab"))}.}

\item{ylab}{Y axis label, same font attributes as \code{xlab}.}

\item{colorPalette}{Vector of color values}

\item{logScale}{logScale}

\item{zTransform}{zTransform}

\item{digits}{Numeric value indicating the number of digits used for numeric values. Whether \code{digitsFunc} is provided, this will be ignored.}

\item{digitsFunc}{Function to be applied to numeric values like doses. This expects a single parameter.}

\item{...}{Further arguments that are passed to \code{\link{format}} function
for formatting of axis labels}
}
\description{
Plot of effect-size object
}
