% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MarginalFit-methods.R
\name{plot.MarginalFit}
\alias{plot.MarginalFit}
\title{Plot monotherapy curve estimates}
\usage{
\method{plot}{MarginalFit}(x, ncol = 2, logScale = TRUE, smooth = TRUE, dataScale = FALSE, ...)
}
\arguments{
\item{x}{Output of \code{\link{fitMarginals}} function or a
\code{"MarginalFit"} object}

\item{ncol}{Number of plots per row}

\item{logScale}{Whether x-axis should be plotted on a logarithmic scale}

\item{smooth}{Whether to draw a smooth fitted curve (deafult), or 
line segments connecting predicted points only}

\item{dataScale}{Whether to draw plot on original data scale in case when 
transformations were used for fitting. Default (FALSE) is to plot on the 
\code{coef(x)} scale}

\item{...}{Further arguments}
}
\value{
Returns a \code{ggplot} object. It can be consequently modified by
  using standard operations on \code{ggplot} objects (if \code{ggplot2}
  package is loaded).
}
\description{
Plot monotherapy curve estimates
}
