% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meanR-functions.R
\name{plot.meanR}
\alias{plot.meanR}
\title{Plot bootstrapped cumulative distribution function of meanR null distribution}
\usage{
\method{plot}{meanR}(x, ...)
}
\arguments{
\item{x}{Output from \code{\link{meanR}}}

\item{...}{Further arguments}
}
\description{
Plot bootstrapped cumulative distribution function of meanR null distribution
}
