% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confInt-functions.R
\name{plotConfInt}
\alias{plotConfInt}
\title{Plot confidence intervals from BIGL object in a contour plot}
\usage{
plotConfInt(BIGLobj, ...)
}
\arguments{
\item{BIGLobj}{Output from \code{\link{fitSurface}}}

\item{...}{passed on to \code{\link{plot.BIGLconfInt}}}
}
\description{
Plot confidence intervals from BIGL object in a contour plot
}
