% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MarginalFit-methods.R
\name{predict.MarginalFit}
\alias{predict.MarginalFit}
\title{Predict values on the dose-response curve}
\usage{
\method{predict}{MarginalFit}(object, newdata, ...)
}
\arguments{
\item{object}{Output of \code{\link{fitMarginals}} function}

\item{newdata}{An optional data frame in which to look for \code{d1} and
\code{d2} variables with which to predict. If omitted, the fitted values
are used. Doses that are passed to this function must correspond to
marginal data, i.e. at least one of the doses must be zero.}

\item{...}{Further arguments}
}
\description{
Predict values on the dose-response curve
}
