% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaleResids.R
\name{predictVar}
\alias{predictVar}
\title{Predict variance}
\usage{
predictVar(means, model)
}
\arguments{
\item{means}{a vector of means}

\item{model}{coefficients of a fitted mean-variance trend}
}
\description{
Predict variance
}
