% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confInt-functions.R
\name{print.summary.BIGLconfInt}
\alias{print.summary.BIGLconfInt}
\title{Print summary of BIGLconfInt object}
\usage{
\method{print}{summary.BIGLconfInt}(x, ...)
}
\arguments{
\item{x}{Summary of BIGLconfInt object}

\item{...}{Further arguments}
}
\description{
Print summary of BIGLconfInt object
}
