% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MarginalFit-methods.R
\name{print.summary.MarginalFit}
\alias{print.summary.MarginalFit}
\title{Print method for summary of \code{MarginalFit} object}
\usage{
\method{print}{summary.MarginalFit}(x, ...)
}
\arguments{
\item{x}{Summary of \code{MarginalFit} object}

\item{...}{Further arguments}
}
\description{
Print method for summary of \code{MarginalFit} object
}
