% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxR-functions.R
\name{print.summary.maxR}
\alias{print.summary.maxR}
\title{Print summary of maxR object}
\usage{
\method{print}{summary.maxR}(x, ...)
}
\arguments{
\item{x}{Summary of \code{"maxR"} object}

\item{...}{Further arguments}
}
\description{
Print summary of maxR object
}
