% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meanR-functions.R
\name{print.summary.meanR}
\alias{print.summary.meanR}
\title{Print summary of meanR object}
\usage{
\method{print}{summary.meanR}(x, ...)
}
\arguments{
\item{x}{Summary of meanR object}

\item{...}{Further arguments}
}
\description{
Print summary of meanR object
}
