% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ResponseSurface-methods.R
\name{print.summary.ResponseSurface}
\alias{print.summary.ResponseSurface}
\title{Print method for the summary function of \code{ResponseSurface} object}
\usage{
\method{print}{summary.ResponseSurface}(x, ...)
}
\arguments{
\item{x}{Summary of \code{ResponseSurface} object}

\item{...}{Further parameters}
}
\description{
Print method for the summary function of \code{ResponseSurface} object
}
