% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MarginalFit-methods.R
\name{residuals.MarginalFit}
\alias{residuals.MarginalFit}
\title{Residuals from marginal model estimation}
\usage{
\method{residuals}{MarginalFit}(object, ...)
}
\arguments{
\item{object}{Output of \code{\link{fitMarginals}} function}

\item{...}{Further arguments}
}
\description{
Residuals from marginal model estimation
}
