% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runBIGL.R
\name{runBIGL}
\alias{runBIGL}
\title{Run the BIGL application for demonstrating response surfaces}
\usage{
runBIGL(...)
}
\arguments{
\item{...}{Pass parameters to \code{\link[shiny]{runApp}}}
}
\description{
Run the BIGL application for demonstrating response surfaces
}
\examples{
\dontrun{
  runBIGL()
}
}
