% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaleResids.R
\name{scaleResids}
\alias{scaleResids}
\title{Functions for scaling, and rescaling residuals. May lead to unstable behaviour in practice}
\usage{
scaleResids(sampling_errors, means, model)
}
\arguments{
\item{sampling_errors}{A vector of raw  residuals}

\item{means}{a vector of means}

\item{model}{coefficients of a fitted mean-variance trend}
}
\description{
Functions for scaling, and rescaling residuals. May lead to unstable behaviour in practice
}
\details{
Residuals are calculated with respect to the average observation on
the off-axis point, so replicates are required!
}
