% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confInt-functions.R
\name{summary.BIGLconfInt}
\alias{summary.BIGLconfInt}
\title{Summary of confidence intervals object}
\usage{
\method{summary}{BIGLconfInt}(object, ...)
}
\arguments{
\item{object}{Output from \code{\link{bootConfInt}}}

\item{...}{Further arguments}
}
\description{
Summary of confidence intervals object
}
