% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MarginalFit-methods.R
\name{summary.MarginalFit}
\alias{summary.MarginalFit}
\title{Summary of \code{MarginalFit} object}
\usage{
\method{summary}{MarginalFit}(object, ...)
}
\arguments{
\item{object}{Output of \code{\link{fitMarginals}} function}

\item{...}{Further arguments}
}
\description{
Summary of \code{MarginalFit} object
}
