% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxR-functions.R
\name{summary.maxR}
\alias{summary.maxR}
\title{Summary of maxR object}
\usage{
\method{summary}{maxR}(object, ...)
}
\arguments{
\item{object}{Object of \code{"maxR"} class}

\item{...}{Further arguments}
}
\description{
Summary of maxR object
}
