% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meanR-functions.R
\name{summary.meanR}
\alias{summary.meanR}
\title{Summary of meanR object}
\usage{
\method{summary}{meanR}(object, ...)
}
\arguments{
\item{object}{Output from \code{\link{meanR}}}

\item{...}{Further arguments}
}
\description{
Summary of meanR object
}
