% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ResponseSurface-methods.R
\name{summary.ResponseSurface}
\alias{summary.ResponseSurface}
\title{Summary of \code{ResponseSurface} object}
\usage{
\method{summary}{ResponseSurface}(object, ...)
}
\arguments{
\item{object}{Output of \code{\link{fitSurface}}}

\item{...}{Further parameters}
}
\description{
Summary of \code{ResponseSurface} object
}
