% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MarginalFit-methods.R
\name{vcov.MarginalFit}
\alias{vcov.MarginalFit}
\title{Estimate of coefficient variance-covariance matrix}
\usage{
\method{vcov}{MarginalFit}(object, ...)
}
\arguments{
\item{object}{Output of \code{\link{fitMarginals}} function}

\item{...}{Further arguments}
}
\description{
Estimate of coefficient variance-covariance matrix
}
