% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaleResids.R
\name{addResids}
\alias{addResids}
\title{Add residuals by adding to mean effects}
\usage{
addResids(means, sampling_errors, method, model)
}
\arguments{
\item{means}{a vector of means}

\item{sampling_errors}{A vector of raw  residuals}

\item{method}{a variance method, see \code{\link{fitSurface}}}

\item{model}{coefficients of a fitted mean-variance trend}
}
\description{
Add residuals by adding to mean effects
}
