% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaleResids.R
\name{addResids}
\alias{addResids}
\title{Add residuals by adding to mean effects}
\usage{
addResids(means, ...)
}
\arguments{
\item{means}{a vector of means}

\item{...}{passed on to predictVar}
}
\description{
Add residuals by adding to mean effects
}
