% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaleResids.R
\name{backscaleResids}
\alias{backscaleResids}
\title{Backscale residuals}
\usage{
backscaleResids(scaledResids, ...)
}
\arguments{
\item{scaledResids}{scaled residuals}

\item{...}{passed on to predictVar}
}
\description{
Backscale residuals
}
