% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ResponseSurface-methods.R
\name{contour.ResponseSurface}
\alias{contour.ResponseSurface}
\title{Method for plotting of contours based on maxR statistics}
\usage{
\method{contour}{ResponseSurface}(x, ...)
}
\arguments{
\item{x}{Output of \code{\link{fitSurface}}}

\item{...}{Further parameters passed to \code{\link{plot.maxR}}}
}
\description{
Method for plotting of contours based on maxR statistics
}
