% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaleResids.R
\name{predictVar}
\alias{predictVar}
\title{Predict variance}
\usage{
predictVar(means, model, invTransFun)
}
\arguments{
\item{means}{a vector of means}

\item{model}{The mean-variance model}

\item{invTransFun}{the inverse transformation function, back to the variance domain}
}
\description{
Predict variance
}
