% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaleResids.R
\name{sampleResids}
\alias{sampleResids}
\title{Sample residuals according to a new model}
\usage{
sampleResids(means, sampling_errors, method, rescaleResids, ...)
}
\arguments{
\item{means}{a vector of means}

\item{sampling_errors}{A vector of raw  residuals}

\item{...}{passed on to predictVar}
}
\value{
sampled residuals
}
\description{
Sample residuals according to a new model
}
